(wip) Extract and encode spans data
Refactoring keyword extraction to extract spans information.

Modifying the intermediate storage of converted data to use the new slop library, which is allows for easier storage of ad-hoc binary data like spans and positions.

This is a bit of a katamari damacy commit that ended up dragging along a bunch of other fairly tangentially related changes that are hard to break out into separate commits after the fact.  Will push as-is to get back to being able to do more isolated work.

+++ code/features-convert/keyword-extraction/java/nu/marginalia/keyword/KeywordMetadata.java
@@ -27,9 +27,9 @@ class KeywordMetadata {
         this.urlKeywords = urlKeywords;
     }
 
-    public long getMetadataForWord(String stemmed) {
+    public byte getMetadataForWord(String stemmed) {
 
-        long flags = 0;
+        byte flags = 0;
 
         if (subjectLikeKeywords.contains(stemmed)) {
             flags |= WordFlags.Subjects.asBit();
